add graph prototype
we are updating the compatibility schema prototypes to be
entirely graphs, meaning that every node must be represented
in the graph. This also makes our lives simpler because we
do not need a separate json schema for some custom format -
we can just use JGF! While there are tools that are using V1,
I reviewed the differences and think V2 is better in that
we can lookup specific nodes in N1 time (without needing
to iterate over a list) and this seems to be the preference
for the json graph maintainers, so we should go that route
and (eventually) update underlying tools that might use
this to support that. In addition, I am adding an example
for a compatibility spec, or what would be an artifact
that represents an application or container image, and
uses one or more compatibility schemas.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>