feat: add support for ior
This addition is a WIP for IOR, which will allow
us to study I/O intents. This currently just includes
the different modules and command line options, and my
next step is to design a library (likely I will use the
namesake "compspec" that I made for graph comparison but
am not currently using) that can serve as a Python-based
extraction tool. I started to write something in Go,
but realized that the scientific community is heavily
Python based and I need to support that.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>