refactor: separating supercontainers into resource types
Problem: we want to be able to ask for subsystem resources on
a more granular level.
Solution: separate supercontainers out into each of mpi, hardware,
network (communication) and os. The use case is that if we are
registering different subsystems to a graph, the different resource
types should be more modularity defined.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>